Separate implementations now compile
